made nbt compounds use hashmaps instead of Vecs, to match vanilla
